final commits before documentation

+++ public/scripts/Controllers/DashboardController.js
@@ -1,45 +1,22 @@
 angular.module('IlLumenNote').controller('DashboardController', function ($filter, $timeout, $state, $scope, $rootScope, LoginService, NotificationService, NoteService) {
   $scope.notificationBody = ' ';
+  NotificationService.$$element = angular.element('.notification').first();
 
-  $scope.notes = [];
+  window.notes = $scope.notes = [
+    {
+      note_title: 'Loading notes',
+      name: 'IlLumenNote',
+      note_body: 'One moment please...',
+      updated_at: '...'
+    }
+  ];
+
   $scope.openNotes = null;
-  $scope.gridElement = angular.element('.grid')[0];
-
-  $scope.reGrid = function () {
-    $timeout(
-      function () {
-        $scope.grid = new Muuri(
-          $scope.gridElement, 
-          {
-            layout: {
-              fillGaps: true
-            },
-            dragEnabled: true, 
-            dragSortPredicate: {
-              action: 'swap', 
-              tolerance: 30
-            }
-          }
-        );
-
-        $scope.grid.on('dragEnd', function (a,b) {
-          localStorage.setItem('$slotConfig', JSON.stringify($scope.grid._layout.slots));
-        });
-
-        var $config = localStorage.getItem('$slotConfig');
-        if ($config) {
-          $scope.grid._layout.slots = JSON.parse($config);
-          $scope.grid.refreshSortData();
-        }
-        console.log('$config', $config)
-      }
-    );
-  };
 
   $scope.alertUserInput = function () {
     NotificationService.color("red");
     NotificationService.notify(
-      function () { $scope.notificationBody = 'Email and password are required to login!'; $scope.$apply(); },
+      function () { $scope.notificationBody = 'There was an error completing the note action!'; $scope.$apply(); },
       function () { $scope.notificationBody = ' '; }
     );
   };
@@ -48,74 +25,107 @@ angular.module('IlLumenNote').controller('DashboardController', function ($filte
     function (success) { 
       $scope.notes = success.data;
       if ($scope.$$phase != '$digest') { $scope.$apply(); }
-      $scope.reGrid();
     },
     $scope.alertUserInput
   );
 
+  $scope.initElement = function (a, b) {
+    // console.log('initElement a', a, b);
+  };
+
   $scope.newNote = function () {
     var newDate = $filter('date')((new Date()), 'yyyy-MM-dd HH:mm:ss');
-    var note = { 
+    var note = {
       owner_id: $rootScope.$user.id,
       name: $rootScope.$user.name, 
       note_body: '...',
       created_at: newDate, 
       note_title: 'New note' 
-    }
+    };
     var length = $scope.notes.push(note);
     $scope.edit(length);
-    $timeout(function () { $scope.grid.add(angular.element('.note-' + length)[0]) });
   };
 
   $scope.getNoteId = function ($index) {
-    return "note-" + ($index + 1);
+    $note_id = $scope.notes[$index].note_id;
+
+    if ($note_id) {
+      return 'note-' + ($scope.notes[$index].note_id) + ' note-index-' + $index;
+    } else {
+      return 'note-index-' + $index;
+    }
   };
 
   $scope.edit = function ($selectedIndex) {
     $scope.notes.forEach(function (each, noteIndex) {
       switch (true) {
         case (each.editting) && !((each.note_id == "") || (each.note_id == null)):
-          console.log('Save note Number ' + noteIndex);
+          each.saving = true;
           NoteService.putNote(each.note_body, each.note_title, each.note_id).then(
             function (success) {
-              console.log('put success', success);
-              $scope.notes[$selectedIndex].note_title = success.data.note_title;
-              $scope.notes[$selectedIndex].note_body = success.data.note_body;
+              each.note_title = success.data.note_title;
+              each.note_body = success.data.note_body;
+              each.updated_at = success.data.updated_at;
+              each.created_at = success.data.created_at;
+              each.saving = false;
+              each.editting = false;
               if ($scope.$$phase != '$digest') { $scope.$apply(); }
             },
-            console.error
+            function (error) {
+              each.saving = false;
+              each.editting = false;
+              console.error(error);
+            }
           );
-          each.editting = false;
           break;
         case (each.editting) && ((each.note_id == "") || (each.note_id == null)):
-          console.log('Save note Number ' + noteIndex);
+          each.saving = true;
           NoteService.postNote(each.note_body, each.note_title).then(
             function (success) {
-              console.log('post success', success);
-              $scope.notes[$selectedIndex].note_title = success.data.note_title;
-              $scope.notes[$selectedIndex].note_body = success.data.note_body;
+              each.note_id = success.data.note_id;
+              each.note_title = success.data.note_title;
+              each.note_body = success.data.note_body;
+              each.updated_at = success.data.updated_at;
+              each.created_at = success.data.created_at;
+              each.saving = false;
+              each.editting = false;
               if ($scope.$$phase != '$digest') { $scope.$apply(); }
             },
-            console.error
+            function (error) {
+              each.saving = false;
+              each.editting = false;
+              console.error(error);
+            }
           );
-          each.editting = false;
           break;
         case ((noteIndex == $selectedIndex) && (!each.editting)):
-          console.log('Begin editting. Break.');
           each.editting = true;
           break;
         default:
+          each.editting = false;
+          each.saving = false;
           break;
       }
     });
   };
 
-  $scope.delete = function ($note_id, $selectedIndex) {
+  $scope.delete = function ($note_id, $selectedIndex, $event) {
+    if (!$note_id) {
+      console.log($selectedIndex);
+      $scope.notes.splice($selectedIndex, 1);
+      return; 
+    }
+    $scope.notes[$selectedIndex].deleting = true;
+    $scope.notes[$selectedIndex].editting = false;
+    $scope.notes[$selectedIndex].saving = false;
     NoteService.deleteNote($note_id).then(
       function (success) {
         $scope.notes.splice($selectedIndex, 1)
       },
-      console.error
+      function (error) {
+        $scope.notes[$selectedIndex].deleting = false;
+        console.error(error);
+      }
     );
   };
